third_party/blink/web_tests/platform/generic/http/tests/eventsource/workers/eventsource-eof-expected.txt

[Worker] Test that EventSource discards event data if there is no newline before eof.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: ../script-tests/eventsource-eof.js
PASS [Worker] evt.type is "message"
PASS [Worker] evt.data is "DATA1"
PASS [Worker] evt.lastEventId is "1"
PASS [Worker] evt.type is "message"
PASS [Worker] evt.data is "DATA2"
PASS [Worker] evt.lastEventId is "2"
PASS [Worker] evt.type is "message"
PASS [Worker] evt.data is "DATA3"
PASS [Worker] evt.lastEventId is "3"
PASS [Worker] evt.type is "msg"
PASS [Worker] evt.data is "DATA"
PASS [Worker] evt.lastEventId is "3.1"
PASS successfullyParsed is true

TEST COMPLETE
